Use Double for floating-point custom data and criteria parameters (#302)
* Use Double for floating-point custom data and criteria parameters

* Fix whitespace

+++ Sources/ApptentiveKit/Apptentive+Compatibility.swift
@@ -376,8 +376,8 @@ extension Apptentive {
                 case let int as Int:
                     result[key] = int
 
-                case let float as Float:
-                    result[key] = float
+                case let double as Double:
+                    result[key] = double
 
                 case let string as String:
                     result[key] = string
